Remove strict separation into nixos and home-manager module files
This had been a point of friction for a while but now it's over. Now the file
are managed together, and all home-manager specific stuff simply is prefixed
with the home-manager.users.username boilerplate. If that turns out to be too
annoying to write, maybe some abstraction can be made. For now, it's much
better.