Commit host configurations to git
This obviates the need for local install branches, making it possible to view
and edit all of the system configurations in one place. Additionally, we gain
a better place to manage host-specific overrides.

Additionally, the cli.sh script is officially out and replaced by a Makefile.
The ergonomics are way better with built-in command completion and dependency
relationship management.